requirements.txt
# InsightAI — Analyst Workstation
# Core
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0

# AI
google-generativeai>=0.7.0

# Visualisation (Power BI–style)
plotly>=5.20.0
kaleido>=0.2.1        # for chart export

# Database
psycopg2-binary>=2.9.9
duckdb>=0.9.0        # in-memory SQL on uploaded dataframes
sqlalchemy>=2.0.0

# Data analyst tools
scipy>=1.12.0         # stats: skewness, normality tests
scikit-learn>=1.4.0   # preprocessing, clustering
openpyxl>=3.1.0       # Excel read/write
xlrd>=2.0.1           # legacy .xls
pyarrow>=15.0.0       # fast parquet/feather I/O